+++ packages/terra-site/src/examples/button-group/ButtonGroupWithIcons.jsx
@@ -1,32 +1,25 @@
 import React from 'react';
-import IconEdit from 'terra-icon/lib/icon/IconEdit';
-import IconAdd from 'terra-icon/lib/icon/IconAdd';
+import IconCaretUp from 'terra-icon/lib/icon/IconCaretUp';
+import IconCaretDown from 'terra-icon/lib/icon/IconCaretDown';
 import ButtonGroup from 'terra-button-group';
 
-const iconEdit = <IconEdit />;
-const iconAdd = <IconAdd />;
+const iconCaretUp = <IconCaretUp />;
+const iconCaretDown = <IconCaretDown />;
 
 const ButtonGroupWithIcons = () => (
   <div>
     <div>
-      <ButtonGroup
-        buttons={[<ButtonGroup.Button text="Icon" icon={iconEdit} key="iconDefault1" />,
-          <ButtonGroup.Button text="Icon" icon={iconAdd} key="iconDefault2" />]}
-      />
+      <ButtonGroup>
+        <ButtonGroup.Button text="Up" key="upButton" />
+        <ButtonGroup.Button text="Down" key="downButton" />
+      </ButtonGroup>
     </div>
     <br />
     <div>
-      <ButtonGroup
-        buttons={[<ButtonGroup.Button text="Icon Reversed" icon={iconEdit} isReversed key="iconReversed1" />,
-          <ButtonGroup.Button text="Icon Reversed" icon={iconAdd} isReversed key="iconReversed2" />]}
-      />
-    </div>
-    <br />
-    <div>
-      <ButtonGroup
-        buttons={[<ButtonGroup.Button icon={iconEdit} key="iconOnly1" />,
-          <ButtonGroup.Button icon={iconAdd} key="iconOnly2" />]}
-      />
+      <ButtonGroup>
+        <ButtonGroup.Button text="Up" icon={iconCaretUp} key="upIcon" />
+        <ButtonGroup.Button text="Down" icon={iconCaretDown} key="downIcon" />
+      </ButtonGroup>
     </div>
   </div>
 );