+++ packages/terra-slide-group/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Update to React 16
 
+### Fixed
+* Slide visibility is maintained when multiple items are added/removed at the same time.
+
 1.16.0 - (February 1, 2018)
 ------------------
 ### Changed